commands: ec2 bundle vol does not bundle disk image anymore
So now ec2-bundle-vol can be run only on live machine (inside ec2) to
bundle the machine itself. It can not be used anymore on disk image.
For that there is now ec2-bundle-image:

-/opt/ec2/bin/ec2-bundle-vol
+/opt/ec2/bin/ec2-bundle-image

So I had to re-write alsmost completely the rocks create bundle since
now the logic has changed a lot. Almost a day of debugging to figure
all this out....

EC2 I love you!!